pyproject.toml
[project]
name = "article-publisher"
version = "0.1.0"
description = "Multi-platform article publishing system"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "python-frontmatter>=1.0.0",
    "jinja2>=3.1.0",
    "httpx>=0.27.0",
    "playwright>=1.40.0",
    "apscheduler>=3.10.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[project.scripts]
publisher = "src.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
